+++ spec/roo_on_rails/routemaster/publisher_spec.rb
@@ -1,10 +1,10 @@
 require 'active_support/core_ext/string'
 require 'roo_on_rails/routemaster/publisher'
+require 'support/test_model'
 
 RSpec.describe RooOnRails::Routemaster::Publisher do
   TestPublisherA = Class.new(RooOnRails::Routemaster::Publisher)
   TestPublisherB = Class.new(RooOnRails::Routemaster::Publisher)
-  TestModel = Class.new
   let(:model) { TestModel.new }
   let(:event) { :noop }
 
@@ -27,10 +27,14 @@ RSpec.describe RooOnRails::Routemaster::Publisher do
         :noop,
         "test_models",
         "https://deliveroo.test/url",
-        { data: {
-          "test_key_1" => "Test value 1",
-          "test_key_2" => "Test value 2"
-        }}
+        {
+          async: false,
+          data: {
+            "test_key_1" => "Test value 1",
+            "test_key_2" => "Test value 2"
+          },
+          t: nil
+        }
       )
       publisher.publish!
     end
@@ -43,11 +47,61 @@ RSpec.describe RooOnRails::Routemaster::Publisher do
       expect(publisher.url).to eq("https://deliveroo.test/url")
     end
 
+    it 'should default to publishing synchronously' do
+      expect(publisher).to_not be_async
+    end
+
     it 'should have the correct event type' do
       expect(publisher.created?).to eq(false)
       expect(publisher.updated?).to eq(false)
       expect(publisher.deleted?).to eq(false)
       expect(publisher.noop?).to eq(true)
+    end
+
+    describe 'the timestamp of the event sent to routemaster' do
+      subject(:timestamp) do
+        ts = nil
+        expect(::Routemaster::Client).to receive(:send) { |_, _, _, opts| ts = opts[:t] }
+        publisher.publish!
+        ts
+      end
+
+      context 'when the model was created' do
+        let(:event) { :created }
+
+        context 'when the model responds to created_at' do
+          let(:create_time) { Time.at(12345) }
+          let(:model) { TestModel.which_responds_to(created_at: create_time).new }
+
+          it { should eq create_time.to_i }
+        end
+
+        context 'when the model does not respond to created_at' do
+          # it { should eq nil }
+
+          context 'when the model responds to updated_at' do
+            let(:update_time) { Time.at(23456) }
+            let(:model) { TestModel.which_responds_to(updated_at: update_time).new }
+
+            it { should eq update_time.to_i }
+          end
+        end
+      end
+
+      context 'when the model was updated' do
+        let(:event) { :updated }
+
+        context 'when the model does not respond to updated_at' do
+          it { should eq nil }
+        end
+
+        context 'when the model responds to updated_at' do
+          let(:update_time) { Time.at(34567) }
+          let(:model) { TestModel.which_responds_to(updated_at: update_time).new }
+
+          it { should eq update_time.to_i }
+        end
+      end
     end
   end