+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Features:
   back-and-forthing (#57)
 - Friendlier error message when facing Heroku permission errors (#58)
 - Replaces the Rails logger with a structured logger (#60)
+- Bugfix for supporting non-default publishers after a `reload!` in the rails console (#62)
 
 
 # v1.10.0 (2017-08-11)